Time: 1 ms (100.00%) | Memory: 62.5 MB (20.14%) - LeetSync

+++ 4005-maximum-total-subarray-value-i/maximum-total-subarray-value-i.java
@@ -1,10 +1,10 @@
 class Solution {
     public long maxTotalValue(int[] nums, int k) {
-        long max=Integer.MIN_VALUE;
-        long min=Integer.MAX_VALUE;
+        int max=Integer.MIN_VALUE;
+        int min=Integer.MAX_VALUE;
         for(int num:nums){
-            max=(long)Math.max(max,num);
-            min=(long)Math.min(min,num);
+            max=Math.max(max,num);
+            min=Math.min(min,num);
         }
         return (long)k*(max-min);
     }
